Move DataStream page to be a modal on SiteDetails page
